requirements.txt
PyQt6>=6.4.0
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
google-generativeai>=0.3.0 